Move to the {pkg,product}build toolchain
Recent versions of Xcode do not contain the PackageMaker application (or
helper utility). The combination of pkgbuild to build components and
productbuild to build whole installer packages is the supported way to
create installer packages form the command line in OS X. Another plus is
that with the upcoming introduction of Mountain Lion and increasing
enforcement of signed code, you can sign the installer package if you
wish.